+++ wakamiti-api/src/main/java/es/iti/wakamiti/api/WakamitiConfiguration.java
@@ -140,7 +140,7 @@ public class WakamitiConfiguration implements ConfigContributor<Void> {
                     OUTPUT_FILE_PATH, "wakamiti.json",
                     OUTPUT_FILE_PER_TEST_CASE, Boolean.FALSE.toString(),
                     REPORT_GENERATION, Boolean.TRUE.toString(),
-                    ID_TAG_PATTERN, "ID-(\\w*)",
+                    ID_TAG_PATTERN, "ID([\\w-]+)",
                     INCLUDE_FILTERED_TEST_CASES, "false",
                     REDEFINITION_ENABLED, Boolean.TRUE.toString(),
                     REDEFINITION_DEFINITION_TAG, "definition",