+++ wakamiti-engine/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+- The junit wakamiti configuration is modified to set the default output path of the current project.
+
+### Added
+- Duration data type
+
+
 ## [2.5.0] - 2024-04-16
 
 ### Fixed